Desktop/disaster_portal_final/requirements.txt
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
branca==0.8.0
cachetools==5.5.0
cassidy==0.1.4
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.2.16
django-leaflet==0.31.0
enumb==0.1.5
expo==0.1.2
feedparser==6.0.11
folium==0.17.0
geographiclib==2.0
geojson==3.1.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mimetype==0.1.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
narwhals==1.9.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pip==24.2
protobuf==5.28.2
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
schedule==1.2.2
setuptools==74.1.3
sgmllib3k==1.0.0
six==1.16.0
